Rename files and fixes
