+++ backend/components/gamificationController.js
